nice to have pagination task completed and default movies added on page load
