+++ index.html
@@ -13,53 +13,6 @@
     <link rel="stylesheet" href="./public/css/style.css" />
   </head>
   <body>
-    <!-- <header>
-      <div class="logo"><img src="./public/img/logo.png" alt="logo" /></div>
-      <p>FeaWatch is the most famous movie platform in the world</p>
-      <a href="#">Register/Sign In</a>
-    </header>
-    <div class="container">
-      <div class="search-bar">
-        <input
-          id="search-input"
-          type="text"
-          placeholder="Search movie with title"
-        />
-        <select id="search-by-category">
-          <option value=""></option>
-          <option value="movie">Movie</option>
-          <option value="series">Series</option>
-          <option value="episode">Episode</option>
-        </select>
-        <select id="dateSelect">
-       
-        </select>
-        <button id="search-button">Search</button>
-      </div>
-      <div class="results">
-        <div class="movie-card">
-          <img
-            class="movie-img"
-            src="./public/img/default-card-img.jpg"
-            alt=""
-          />
-          <div class="movie-information">
-            <h1>Movie Title</h1>
-            <p>Lorem ipsum dolor sit amet consectetur Lorem ipsum dolor sit</p>
-          </div>
-        </div>
-      </div>
-    </div>
-    <footer>
-      <div class="applications">
-        <img src="./public/img/googleplay.png" alt="googleplay" />
-        <img src="./public/img/appstore.png" alt="applestore" />
-      </div>
-      <div class="copy-right">
-        <p>Designed by Ali Demirci @2022</p>
-      </div>
-    </footer> -->
-
     <script type="module" src="./src/init/initApp.js"></script>
   </body>
 </html>
