+++ public/css/_header.scss
@@ -32,5 +32,8 @@ header {
 @media (max-width: 767px) {
   header {
     flex-direction: column;
+    p {
+      display: none;
+    }
   }
 }
