+++ public/css/_pagination.scss
@@ -1,0 +1,18 @@
+#pagination {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  #next-movie,
+  #prev-movie {
+    margin: 1rem;
+    align-items: center;
+    color: $darkRed;
+    transition: all 0.5s;
+    &:hover {
+      text-decoration: underline;
+      color: $lightWhite;
+      cursor: pointer;
+    }
+  }
+}
