+++ public/css/_results.scss
@@ -1,0 +1,44 @@
+.results {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  align-items: center;
+  padding: 2rem;
+  .movie-card {
+    margin: 0.5rem;
+    position: relative;
+    width: 20%;
+    color: $lightWhite;
+    overflow: hidden;
+
+    .movie-img {
+      border: solid 2px $darkRed;
+      width: 97%;
+    }
+    .movie-information {
+      position: absolute;
+      display: flex;
+      flex-direction: column;
+      justify-content: center;
+      align-items: center;
+      padding: 0.3rem;
+      opacity: 0;
+      top: 1%;
+      left: 1%;
+      background-color: $black;
+      width: 93%;
+      height: 95%;
+      transition: all 1s;
+      color: $orangeEx;
+      cursor: pointer;
+    }
+    &:hover {
+      .movie-information {
+        opacity: 0.9;
+      }
+      .movie-img {
+        border: solid 2px $orangeEx;
+      }
+    }
+  }
+}
