+++ public/css/style.scss
@@ -1,1 +1,2 @@
-@import 'colors', 'main', 'header', 'container', 'footer', 'carddetail';
+@import 'colors', 'main', 'header', 'container', 'footer', 'carddetail',
+  'pagination', 'results';
