+++ src/views/createContainer.js
@@ -1,5 +1,5 @@
+import getSearchQueries from '../handlers/getSearchQueries.js';
 import createElement from '../helpers/createElement.js';
-import createCard from './createCard.js';
 
 const createContainer = () => {
   const container = createElement('div', 'container', 'container');
@@ -7,17 +7,12 @@ const createContainer = () => {
   const searchInput = createElement('input', 'search-input');
   searchInput.setAttribute('placeholder', 'Search with title');
   const selectBoxCategory = createElement('select', 'search-by-category');
-  // options Array will be added
   const selectBoxDate = createElement('select', 'dateSelect');
   const searchButton = createElement('button', 'search-button');
   searchButton.textContent = 'Search';
 
   // Result UI Creation Steps
-
   const results = createElement('div', 'results', 'results');
-
-  // Card Component
-  // const card = createCard();
 
   // Executing to UIS
   container.appendChild(searchBar);
@@ -26,7 +21,7 @@ const createContainer = () => {
   searchBar.appendChild(selectBoxCategory);
   searchBar.appendChild(selectBoxDate);
   searchBar.appendChild(searchButton);
-  // results.appendChild(card);
+  searchButton.addEventListener('click', getSearchQueries);
 
   return container;
 };
