--- src/views/createResults.js
@@ -1,16 +1,0 @@
-const createResults = () => {
-  return `  <div class="results">
-  <div class="movie-card">
-    <img
-      class="movie-img"
-      src="./public/img/default-card-img.jpg"
-      alt=""
-    />
-    <div class="movie-information">
-      <h1>Movie Title</h1>
-      <p>Lorem ipsum dolor sit amet consectetur Lorem ipsum dolor sit</p>
-    </div>
-  </div>
-</div>`;
-};
-export default createResults;
