--- src/views/createSearchByCategory.js
@@ -1,9 +1,0 @@
-const createSearchByCategory = () => {
-  return `<select id="search-by-category">
-  <option value=""></option>
-  <option value="movie">Movie</option>
-  <option value="series">Series</option>
-  <option value="episode">Episode</option>
-</select>`;
-};
-export default createSearchByCategory;
